# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: Agent OS API
  version: 0.6.0
  description: >-
    Control plane and observability API for the Agent OS SPOG. Route
    surface is generated authoritatively from the chi router; payload
    shapes are defined by the SQL migrations (internal/db/migrations)
    and the sqlc query layer (internal/db/queries).
servers:
  - url: http://localhost:8080
    description: Local / hpms1 deployment
...
paths:
...
  /api/projects:
    get:
      tags: [projects]
      operationId: get_ListWorkspaces
      summary: List Workspaces
      description: Handler ListWorkspaces in internal/api.
      responses:
        '200': { description: Success }
        '400': { description: Bad request }
        '404': { description: Not found }
        '500': { description: Server error }
    post:
      tags: [projects]
      operationId: post_CreateWorkspace
      summary: Create Workspace
      description: Handler CreateWorkspace in internal/api.
      responses:
        '200': { description: Success }
        '400': { description: Bad request }
        '404': { description: Not found }
        '500': { description: Server error }
  /api/projects/{id}:
    get:
      tags: [projects]
      operationId: get_GetWorkspace
      summary: Get Workspace
      description: Handler GetWorkspace in internal/api.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Success }
        '400': { description: Bad request }
        '404': { description: Not found }
        '500': { description: Server error }
    patch:
      tags: [projects]
      operationId: patch_UpdateWorkspace
      summary: Update Workspace
      description: Handler UpdateWorkspace in internal/api.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Success }
        '400': { description: Bad request }
        '404': { description: Not found }
        '500': { description: Server error }
  /api/projects/{id}/surface:
    get:
      tags: [projects]
      operationId: get_WorkspaceSurface
      summary: Workspace Surface
      description: Handler WorkspaceSurface in internal/api.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Success }
        '400': { description: Bad request }
        '404': { description: Not found }
        '500': { description: Server error }
...